refactor: `generate_core_characters_getter()` を移動
